requirements.txt
Flask==2.2.5
psycopg2-binary==2.9.6
opencv-python==4.8.0.76
Pillow==9.5.0
rembg==2.0.34
torch==2.0.1
torchvision==0.15.2
numpy==1.25.0
pytest==7.4.0
requests==2.31.0

